src/pyproject.toml
[tool.poetry]
name = "nba-records-rag-chatbot"
version = "0.1.0"
description = "A NBA Chatbot who uses rag to return records about the league"
authors = ["Maximiliano Meyer <maximilianomeyer48@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "~3.10"
numpy = "1.26.4"
openai = "1.60.1"
langchain = "0.3.15"
langchain-openai = "0.3.2"
langchain-google-firestore = "0.5.0"
pinecone-client = "5.0.1"
fastapi = "0.115.7"
uvicorn = "0.34.0"
RapidFuzz = "3.11.0"
firestore = "0.0.8"
firecrawl-py = "1.12.0"
redis = "5.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
